requirements.txt
streamlit 
openai
langchain-openai
langchain-core
python-dotenv
langchain_community
faiss-cpu
ipykernel
datasets
google-generativeai
langchain_openai